% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEveMapLabels}
\alias{getEveMapLabels}
\title{Get the Structure Labels for an Eve Segmentation}
\usage{
getEveMapLabels(type = c("I", "II", "III"))
}
\arguments{
\item{type}{Version of table/map to use}
}
\value{
\code{data.frame} of labels and indices
}
\description{
Returns the structure labels for the different look up table 
images/maps from an Eve segmentation
}
\examples{
getEveMapLabels()
}

