% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEveMapPath}
\alias{getEveMapPath}
\title{Get Eve Path of Look up Table/Map Image}
\usage{
getEveMapPath(type = c("I", "II", "III"))
}
\arguments{
\item{type}{Version of table/map to use}
}
\value{
Character filename of path
}
\description{
Returns the path of a Eve look up table (LUT) image/map.
}
\examples{
getEveMapPath()
}

