% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEvePath}
\alias{getEvePath}
\title{Get Eve Path of File}
\usage{
getEvePath(what = c("T1", "T2", "Brain", "Brain_Mask"))
}
\arguments{
\item{what}{Type of image to return}
}
\value{
Character filename of path
}
\description{
Returns the path of a MNI file to the user.
}
\examples{
getEvePath()
}

