% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEveSegPath}
\alias{getEveSegPath}
\title{Get Eve Path of Segmentation File}
\usage{
getEveSegPath(alg = "FAST")
}
\arguments{
\item{alg}{Algorithm of the segmentation}
}
\value{
Character filename of path
}
\description{
Returns the path of a Eve file of a segmentation to the user.
}
\examples{
getEveSegPath()
}

