% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readEve}
\alias{readEve}
\title{Reads Eve File into R}
\usage{
readEve(..., reorient = FALSE)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{getEvePath}}}

\item{reorient}{Should the image be reoriented? 
Passed to \code{\link{readNIfTI}}}
}
\value{
Object of class \code{nifti}
}
\description{
Matches the argument to pass to \code{\link{getEvePath}},
then reads the NIfTI image into R
}
\examples{
readEve(what = "Brain_Mask")
}

