% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readEveMap}
\alias{readEveMap}
\title{Reads Eve Map File into R}
\usage{
readEveMap(..., reorient = FALSE)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{getEveMapPath}}}

\item{reorient}{Should the image be reoriented? 
Passed to \code{\link{readNIfTI}}}
}
\value{
Object of class \code{nifti}
}
\description{
Matches the argument to pass to \code{\link{getEveMapPath}},
then reads the NIfTI image into R
}
\examples{
readEveMap()
}

