% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readEveSeg}
\alias{readEveSeg}
\title{Reads Eve Segmentation File into R}
\usage{
readEveSeg(..., reorient = FALSE, verbose = TRUE)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{getEveSegPath}}}

\item{reorient}{Should the image be reoriented? 
Passed to \code{\link{readNIfTI}}}

\item{verbose}{Print diagnostic messages of the labels}
}
\value{
Object of class \code{nifti}
}
\description{
Matches the argument to pass to \code{\link{getEveSegPath}},
then reads the NIfTI image of the segmetation into R
}
\examples{
readEveSeg()
}

